Update Karel.md
Karel is a robot created by a student at Stanford University. Karel travels India for the first time and it travels the northern part of India first picking up a beeper each time it comes across a monkey. Then it travels to the southern part of India and visits temples and picks up a beeper every time it rings a bell. As Karel travels around India, it paints each corner a color from the Indian flag (Orange, White, Green and Blue for the ashoka chakra). Karel is in a 10x10 world.